chore: tidy onCreateNode
